// lib/api/generated/models/pageObjectOfMessageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { MessageResponse } from './messageResponse';
import type { PageableObject } from './pageableObject';
import type { SortObject } from './sortObject';

export interface PageObjectOfMessageResponse {
  content?: MessageResponse[];
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}